% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_2columns.R
\name{clean_2columns}
\alias{clean_2columns}
\title{clean_2columns}
\usage{
clean_2columns(
  df,
  col1,
  col2,
  clean = TRUE,
  omit_stops = TRUE,
  lemmatize = TRUE
)
}
\arguments{
\item{df}{a dataframe with two columns of words you want pairwise distance for}

\item{col1}{quoted column name storing the first string for comparison}

\item{col2}{quoted column name storing the second string for comparison}

\item{clean}{T/F default is T specifies whether to apply cleaning transformations or leave data alone}

\item{omit_stops}{T/F user wishes to remove stopwords (default is TRUE)}

\item{lemmatize}{T/F user wishes to lemmatize each string (default is TRUE)}
}
\value{
a dataframe
}
\description{
Cleans a transcript where there are two or more talkers. User specifies the dataframe and column name where target text is stored as arguments to the function. Default option is to lemmatize strings. Function splits and unlists text so that the output is in a one-row-per-word format marked by a unique numeric identifier (i.e., 'id_orig')
}
