% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dyad.R
\name{clean_dyad}
\alias{clean_dyad}
\title{clean_dyad}
\usage{
clean_dyad(df, wordcol, whotalks, lemmatize = TRUE)
}
\arguments{
\item{wordcol}{quoted column name storing the strings that will be cleaned and split}

\item{whotalks}{quoted column name with speaker/talker identities will be factorized}

\item{dat}{a dataframe with at least one target column of string data}
}
\value{
a dataframe one-word-per-row format with 'id_orig', 'word_clean', 'talker' vars appended
}
\description{
Cleans a transcript where there are two or more talkers. User specifies the dataframe and column name where target text is stored as arguments to the function. Default option is to lemmatize strings. Function splits and unlists text so that the output is in a one-row-per-word format marked by a unique numeric identifier (i.e., 'id_orig')
}
