% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_unordered4matrix.R
\name{clean_unordered4matrix}
\alias{clean_unordered4matrix}
\title{clean_unordered4matrix}
\usage{
clean_unordered4matrix(
  df,
  wordcol,
  clean = TRUE,
  omit_stops = TRUE,
  lemmatize = TRUE
)
}
\arguments{
\item{df}{a dataframe with at least one target column of string data}

\item{wordcol}{quoted column name storing the strings that will be cleaned and split}

\item{clean}{apply cleaning functions (lowercase etc) default is TRUE}

\item{omit_stops}{option for omitting stopwords default is TRUE}

\item{lemmatize}{option for lemmatizing strings default is TRUE}
}
\value{
a dataframe
}
\description{
Cleans and formats text. User specifies the dataframe and column name where target text is stored. Word order does not matter (all words shuffled later). Cleaning takes only first instance of word.
}
