% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ColumnSample}
\alias{ColumnSample}
\title{Sample Transcript Split by Two Columns}
\format{
## "ColumnSample"
A data frame with 27 rows and 2 columns:
\describe{
  \item{word1}{text corresponding to the first word in a pair to contrast}
  \item{word2}{text corresponding to the second word in a pair to contrast}
  ...
}
}
\usage{
ColumnSample
}
\description{
No talker delineated. User arrays first word in one column, second word in another column of the dataframe.
}
\keyword{datasets}
