% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dialogue_Dirty}
\alias{Dialogue_Dirty}
\title{Dialogue 'Dirty' Transcript}
\format{
## "Dialogue_Dirty"
A data frame with 75 rows and 2 columns:
\describe{
  \item{word}{text from a language transcript}
  \item{speaker}{Mary or Peter: fictional speaker identities}
  ...
}
}
\usage{
Dialogue_Dirty
}
\description{
A sample dyadic conversation transcript where two people are taslking; Messy string data in a conversation tramscript, mulyiple lines per person per turn, missing observations, fragments, punctuations interspersed with single words. Two people talking to each other.
}
\keyword{datasets}
