% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dialogue_Structured}
\alias{Dialogue_Structured}
\title{Dialogue Structured/Ideal Formatted Transcript}
\format{
## "Dialogue_Structured"
A data frame with 50 rows and 2 vars:
\describe{
  \item{mytext}{text from a language transcript}
  \item{speaker}{P1 or P2 fictional interlocutor identities}
  ...
}
}
\usage{
Dialogue_Structured
}
\description{
Perfectly pre-fomrmatted data in a structure where 2 people are coversing in 1-word utterances back and forth.
}
\keyword{datasets}
