% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dialogue_Typical}
\alias{Dialogue_Typical}
\title{A Typical Dialogue Transcript}
\format{
## "Dialogue_Typical"
A data frame with 5 rows and 2 columns:
\describe{
  \item{word}{fictional text from a language transcript}
  \item{speaker}{Mary or Peter: fictional speaker identities}
  ...
}
}
\usage{
Dialogue_Typical
}
\description{
A sample dyadic conversation transcript where two people are conversing.
}
\keyword{datasets}
