% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DialogueSample1}
\alias{DialogueSample1}
\title{A Sample Dialogue Transcript}
\format{
## "DialogueSample1"
A data frame with 75 rows and 2 columns:
\describe{
  \item{word}{text from a language transcript}
  \item{speaker}{Mary or Peter: fictional speaker identities}
  ...
}
}
\usage{
DialogueSample1
}
\description{
Messy string data composed of missing observations, fragments, punctuations interspersed with single words. Two people talking to each other.
}
\keyword{datasets}
