% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_2cols.R
\name{dist_2cols}
\alias{dist_2cols}
\title{dist_2cols}
\usage{
dist_2cols(dat)
}
\arguments{
\item{dat}{a dataframe prepped using clean_2columns' with word pairs arrayed in two columns}
}
\value{
a dataframe
}
\description{
Function takes dataframe cleaned using 'clean_2columns', computes two metrics of semantic distance for each word pair arrayed in Col1 vs. Col2
}
