% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_dialogue.R
\name{dist_dialogue}
\alias{dist_dialogue}
\title{dist_dialogue}
\usage{
dist_dialogue(dat, who_talking)
}
\arguments{
\item{dat}{a dataframe prepped using 'clean_dialogue' fn with talker data and turncount appended}

\item{who_talking}{factor variable with two levels specifying an ID for the person producing the text in `word_clean`}
}
\value{
a dataframe
}
\description{
Function takes dataframe cleaned using 'clean_dialogue' and computes two metrics of semantic distance turn-to-turn indexing a 'talker' column. Sums all the respective semantic vectors within each tuern, cosine distance to the next turn's composite vector
}
