% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_dialogue_turns.R
\name{dist_dialogue_turns}
\alias{dist_dialogue_turns}
\title{dist_dialogue_turns}
\usage{
dist_dialogue_turns(dat)
}
\arguments{
\item{dat}{a dataframe prepped using 'clean_dialogue' fn with talker data and turncount appended}
}
\value{
a dataframe
}
\description{
Function takes dataframe cleaned using 'clean_monologue', computes two metrics of semantic distance for each word relative to the average of the semantic vectors within an n-word window appearing before each word. User specifies the window (ngram) size. The window 'rolls' across the language sample providing distance metrics
}
