% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_matrix_all.R
\name{dist_matrix_all}
\alias{dist_matrix_all}
\title{dist_matrix_all}
\usage{
dist_matrix_all(dat, dist_type = "embedding")
}
\arguments{
\item{dat}{a dataframe prepped using 'clean_unordered4matrix' fn}

\item{dist_type}{semantic norms for running distance matrix on default='embedding', other is 'SD15'}
}
\value{
a dataframe
}
\description{
Function takes dataframe cleaned using 'clean_unordered4matrix', pairwise distance between all elements as a matrix
}
