% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_paired_cols.R
\name{dist_paired_cols}
\alias{dist_paired_cols}
\title{dist_paired_cols}
\usage{
dist_paired_cols(dat)
}
\arguments{
\item{dat}{a dataframe prepped using clean_2columns' with word pairs arrayed in two columns}
}
\value{
a dataframe
}
\description{
Function takes dataframe cleaned using 'clean_2columns', computes two metrics of semantic distance for each word pair arrayed in Col1 vs. Col2
}
