% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FakeCats}
\alias{FakeCats}
\title{Sample Dataframe with Blocks of Words Related by Semantic Category}
\format{
## "FakeCats"
A data frame with 50 rows and 4 columns:
\describe{
  \item{ID_JR}{a sequential numeric identifier}
  \item{word}{target text}
  \item{category}{semantic category of the target word}
  \item{prediction}{is the target word within category or at a switchpoint between cats}
  ...
}
}
\usage{
FakeCats
}
\description{
No talker delineated. Vector of 50 words, 10 from each of 5 categories (animals, fruits, weapons..).
}
\keyword{datasets}
