% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glowca_25}
\alias{glowca_25}
\title{Glove Semantic Embeddings}
\format{
## "glowca_25"
A data frame with 59061 observations of 301 variables
\describe{
  \item{word}{word characterized across embeddings}
  \item{Param_1}{hyperparameter number 1}
  \item{Param_300}{hyperparameter number 300}
  ...
}
}
\usage{
glowca_25
}
\description{
Word embeddings (300 hyperparameter dimensions, 59061 words). Each word is one row.
}
\keyword{datasets}
