% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Grandfather_Passage}
\alias{Grandfather_Passage}
\title{The Grandfather Passage: A Standardized Reading Passage}
\format{
## "Grandfather_Passage"
A data frame with 1 observation of 1 variable:
\describe{
  \item{mytext}{text from the Grandfather Passage unsplit}
  ...
}
}
\usage{
Grandfather_Passage
}
\description{
A monologue discourse sample. Grandfather Passage is a well-known test of reading aloud.
}
\keyword{datasets}
