% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_github_data}
\alias{load_github_data}
\title{Load all .rda files from a GitHub data folder into the package environment}
\usage{
load_github_data(
  repo = "Reilly-ConceptsCognitionLab/SemanticDistance_Data",
  branch = "main",
  data_folder = "data",
  envir = parent.frame()
)
}
\arguments{
\item{repo}{GitHub repository (e.g., "username/repo")}

\item{branch}{Branch name (default: "main")}

\item{data_folder}{Remote folder containing .rda files (default: "data/")}

\item{envir}{Environment to load into (default: package namespace)}
}
\value{
nothing, loads data (as rda files) from github repository needed for other package functions
}
\description{
Load all .rda files from a GitHub data folder into the package environment
}
