% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Monologue_Dirty}
\alias{Monologue_Dirty}
\title{A Sample 'Dirty' Monologue Transcript}
\format{
## "Monologue_Dirty"
A data.frame with 74 obs and 1 var
\describe{
  \item{mytext}{text from a hypothetical language transcript}
  ...
}
}
\usage{
Monologue_Dirty
}
\description{
No talker delineated. Messy string data composed of missing observations, fragments, punctuations interspersed with single words.
}
\keyword{datasets}
