% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Monologue_Typical}
\alias{Monologue_Typical}
\title{A Typical Monologue Transcript}
\format{
## "Monologue_Typical"
A data frame with 1 row and 1 column
\describe{
  \item{mytext}{text from a language transcript}
  ...
}
}
\usage{
Monologue_Typical
}
\description{
Dataframe with ordered text squashed into a single cell.
}
\keyword{datasets}
