% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MonologueSample1}
\alias{MonologueSample1}
\title{A Sample Monologue Transcript}
\format{
## "MonologueSample1"
A data frame with 74 rows and 1 column:
\describe{
  \item{word}{text from a hypothetical language transcript}
  ...
}
}
\usage{
MonologueSample1
}
\description{
No talker delineated. Messy string data composed of missing observations, fragments, punctuations interspersed with single words.
}
\keyword{datasets}
