% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SD15_2025_complete}
\alias{SD15_2025_complete}
\title{SD15_2025_complete Experiential Semantic Distance Values}
\format{
## "SD15_2025_complete"
A data frame with 25,050 observations of 16 variables
\describe{
  \item{word}{word characterized across 15 ratings}
  \item{Param_auditory_z}{z-score of auditory salience from Lancaster Sensorimotor Norms}
  \item{Param_gustatory_z}{z-score of gustatory salience from Lancaster Sensorimotor Norms}
  \item{Param_haptic_z}{z-score of haptic salience from Lancaster Sensorimotor Norms}
  \item{Param_interoceptive_z}{z-score of interoceptive salience from Lancaster Sensorimotor Norms}
  \item{Param_visual_z}{z-score of visual salience from Lancaster Sensorimotor Norms}
  \item{Param_olfactory_z}{z-score of olfactory salience from Lancaster Sensorimotor Norms}
  \item{Param_handarm_z}{z-score of handarm motor salience from Lancaster Sensorimotor Norms}
  \item{Param_excitement_z}{z-score of excitement salience from affectvec}
  \item{Param_surprised_z}{z-score of surprise salience from affectvec}
  \item{Param_fear_z}{z-score of fear salience from affectvec}
  \item{Param_anger_z}{z-score of anger salience from affectvec}
  \item{Param_disgust_z}{z-score of disgust salience from affectvec}
  \item{Param_sadness_z}{z-score of sadness salience from affectvec}
  \item{Param_happiness_z}{z-score of happiness salience from affectvec}
  \item{Param_contempr_z}{z-score of contempt salience from affectvec}
  ...
}
}
\usage{
SD15_2025_complete
}
\description{
Word embeddings (300 dimensions, 59061 words). Each word is one row.
}
\keyword{datasets}
