% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Semantic_Clusters}
\alias{Semantic_Clusters}
\title{Simulated Semantic Category Fluency Data: Word List Blocked by Semantic Category}
\format{
## "Semantic_Clusters"
A data frame with 20 rows and 3 columns:
\describe{
  \item{ID_JR}{a sequential numeric identifier}
  \item{word}{target text}
  \item{category}{semantic category of the target word}
  ...
}
}
\usage{
Semantic_Clusters
}
\description{
No talker delineated. Vector of 20 words, 5 from each of 4 categories, Good for examining clustering
}
\keyword{datasets}
