% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Temple_stops25}
\alias{Temple_stops25}
\title{Stopword List}
\format{
## "Temple_stops25"
A data frame with 829 observations of 4 variables
\describe{
  \item{id_orig}{numeric identifier}
  \item{word}{stopword target}
  \item{length}{length in words}
  \item{pos}{universal part-of-speech tag}
  ...
}
}
\usage{
Temple_stops25
}
\description{
List of stopwords
}
\keyword{datasets}
