% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Unordered_List}
\alias{Unordered_List}
\title{Unordered_List}
\format{
## "Unordered_List"
A data frame with 20 rows and 3 columns:
\describe{
  \item{ID_JR}{a sequential numeric identifier}
  \item{word}{target text}
}
}
\usage{
Unordered_List
}
\description{
No talker delineated. Vector of 20 words, 5 from each of 4 categories, Good for examining clustering
}
\keyword{datasets}
