% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_clusters.R
\name{viz_clusters}
\alias{viz_clusters}
\title{viz_clusters}
\usage{
viz_clusters(distmat, type = "dendrogram")
}
\arguments{
\item{distmat}{a square matrix of cosine dist values (not a distance matrix!) generated by 'dist_matrix' function}

\item{type}{quoted argument 'dendrogram' is default, 'network' is alternative}
}
\value{
a plot of a dendrogram
}
\description{
viz_clusters
}
