% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_dendrogram.R
\name{viz_dendrogram}
\alias{viz_dendrogram}
\title{viz_dendrogram}
\usage{
viz_dendrogram(dat, dist_type = "embedding")
}
\arguments{
\item{dat}{datafrane with cosine semantic dist values (not a distance matrix!)}

\item{dist_type}{quoted argument semantic norms for running distance matrix on default='embedding', other is 'SD15'}

\item{...}{additional arguments passed to internal functions}
}
\value{
a plot of a dendrogram
}
\description{
Takes a vector of words with semantic distance ratings, converts to a square matrix, then to a euclidean distance
matrix (all word pairs), then plots the words s a cluster dendrogram, behind
the scenes optimizing for k-cluster size.
}
\details{
This function internally calls eval_kmeans_clustersize for
cluster evaluation. The dendrogram visualization is based on hierarchical
clustering of semantic distances.
}
