% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_dists_continuous.R
\name{viz_dists_continuous}
\alias{viz_dists_continuous}
\title{viz_dists_continuous}
\usage{
viz_dists_continuous(
  dat,
  lang_smaple = "monologue",
  interpolate = TRUE,
  roll_avg = 0,
  annotate = TRUE
)
}
\arguments{
\item{dat}{dataframe with cosine semantic dist values in columns (CosDist_) labeled by 'row_id_postsplit'(either monologue or dialogue)}

\item{interpolate}{T/F linear interpolation across missing observations of 'row_id_postsplit' (default is T)}

\item{roll_avg}{window for computing a rolling average for smoothing the data, default is 0 (no smoothing)}

\item{annotate}{T/F option for annotating large semantic distance jumps z>2.5 with a red vertical line}

\item{lang_sample}{quoted argument 'monologue' or 'dialogue', default is monologue}
}
\value{
a plot
}
\description{
Produces time varying plots of semantic distance changes by kernel size (ngram, word, turn) at each
sequential timepoint. Annotated large distance jumps (z>2.5) with a red line.
}
