% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_kmeans_clustersize.R
\name{viz_kmeans_clustersize}
\alias{viz_kmeans_clustersize}
\title{viz_kmeans_clustersize}
\usage{
viz_kmeans_clustersize(distmat, k.max = 15, nboot = 150)
}
\arguments{
\item{distmat}{a square matrix of cosine dist values generated by 'dist_matrix' function}

\item{k.max}{maximum number of clusters to evaluate default is 15}

\item{nboot}{number of bootstrap samples to determine gap statistic, default is 150}
}
\value{
optimal_k an integer (K) representing optimal cluster size for distance matrix by gap statistic
}
\description{
viz_kmeans_clustersize
}
