% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_distmatrix2dendrogram.R
\name{viz_matrix2dendrogram}
\alias{viz_matrix2dendrogram}
\alias{viz_monologue}
\title{viz_matrix2dendrogram}
\usage{
viz_monologue(
  dat,
  interpolate = TRUE,
  roll_avg = 0,
  facet = TRUE,
  annotate = TRUE
)
}
\arguments{
\item{distmat}{a square matrix of cosine dist values generated by 'dist_matrix' function}
}
\value{
a plot of a dendrogram
}
\description{
viz_matrix2dendrogram
}
