% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_monologue.R
\name{viz_monologue}
\alias{viz_monologue}
\title{viz_monologue}
\usage{
viz_monologue(
  dat,
  interpolate = TRUE,
  roll_avg = 0,
  facet = TRUE,
  annotate = TRUE
)
}
\arguments{
\item{dat}{a dataframe with semantic distance values in columns (CosDist_) labeled by 'row_id_postsplit'}

\item{interpolate}{T/F linear interpolation across missing observations of 'row_id_postsplit' (default is T)}

\item{roll_avg}{window for computing a rolling average for smoothing the data, default is 0 (no smoothing)}

\item{facet}{T/F facets the data by distance measure (CosDist_Glo vs. CosDist_SD15)}

\item{annotate}{T/F option for annotating large semantic distance jumps z>2.5 with a red vertical line}
}
\value{
a dataframe
}
\description{
viz_monologue
}
