% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_network.R
\name{viz_network}
\alias{viz_network}
\title{viz_network}
\usage{
viz_network(dat, dist_type = "embedding")
}
\arguments{
\item{dat}{dataframe with cosine semantic distance values (not a distance matrix!)}

\item{dist_type}{quoted argument specifying semantic norms for distance matrix.
Default='embedding', other option is 'SD15'}

\item{...}{additional arguments passed to internal functions}
}
\value{
an igraph network plot
}
\description{
Creates a network visualization from semantic distance data
}
\details{
This function internally calls eval_kmeans_clustersize for
         cluster evaluation. The network visualization is based on hierarchical
         clustering of semantic distances.
}
