% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Word_Pairs}
\alias{Word_Pairs}
\title{Word Pairs in Columns}
\format{
## "Word_Pairs"
A data frame with 27 rows and 2 columns:
\describe{
  \item{word1}{text corresponding to the first word in a pair to contrast}
  \item{word2}{text corresponding to the second word in a pair to contrast}
  ...
}
}
\usage{
Word_Pairs
}
\description{
first target word for computing distance in one column, second word in another column.
}
\keyword{datasets}
