% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words_to_network.R
\name{words_to_network}
\alias{words_to_network}
\title{words_to_network}
\usage{
words_to_network(dat, wordcol, output = "dendrogram", dist_type = "embedding")
}
\arguments{
\item{dat}{dataframe with text in it (not cleaned or formatted)}

\item{wordcol}{quoted argument identifying column in dataframe with target text}

\item{output}{quoted argument for type of output default is 'dendrogram', alternate is 'network'}

\item{dist_type}{quoted argument semantic norms for running distance matrix on default='embedding', other is 'SD15'}

\item{...}{additional arguments passed to internal functions}
}
\value{
a plot of a dendrogram or an igraph network AND a cosine distance matrix
}
\description{
Takes a vector of words with semantic distance ratings, converts to a square matrix, then to a euclidean distance
matrix (all word pairs), then plots the words in either a cluster dendrogram or simple igraph network
}
\details{
This function internally calls eval_kmeans_clustersize for
cluster evaluation. The dendrogram visualization is based on hierarchical
clustering of semantic distances.
}
