% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_2cols.R
\name{clean_2cols}
\alias{clean_2cols}
\title{clean_2cols}
\usage{
clean_2cols(df, col1, col2, clean = TRUE, omit_stops = TRUE, lemmatize = TRUE)
}
\arguments{
\item{df}{a dataframe with two columns of words you want pairwise distance for}

\item{col1}{quoted column name storing the first string for comparison}

\item{col2}{quoted column name storing the second string for comparison}

\item{clean}{T/F default is T specifies whether to apply cleaning transformations or leave data alone}

\item{omit_stops}{T/F user wishes to remove stopwords (default is TRUE)}

\item{lemmatize}{T/F user wishes to lemmatize each string (default is TRUE)}
}
\value{
a dataframe
}
\description{
Cleans a transcript where word pairs are arrayed in two columns.
}
