% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dialogue.R
\name{clean_dialogue}
\alias{clean_dialogue}
\title{clean_dialogue}
\usage{
clean_dialogue(
  df,
  wordcol,
  whotalks,
  clean = TRUE,
  omit_stops = TRUE,
  lemmatize = TRUE,
  split_stings = TRUE
)
}
\arguments{
\item{df}{a dataframe with at least one target column of string data}

\item{wordcol}{quoted column name storing the strings that will be cleaned and split}

\item{whotalks}{quoted column name with speaker/talker identities will be factorized}

\item{clean}{T/F apply cleaning transformations (default is TRUE)}

\item{omit_stops}{T/F user wishes to remove stopwords (default is TRUE)}

\item{lemmatize}{T/F user wishes to lemmatize each string (default is TRUE)}

\item{split_strings}{T/F (default is TRUE) will split and unlist contents of rows}
}
\value{
a dataframe
}
\description{
Cleans a transcript where there are two or more talkers. User specifies the dataframe and column name where target text is stored as arguments to the function. Default option is to lemmatize strings. Function splits and unlists text so that the output is in a one-row-per-word format marked by a unique numeric identifier (i.e., 'id_orig')
}
