% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Unordered_List}
\alias{Unordered_List}
\title{Unordered_List}
\format{
## "Unordered_List"
A data frame with 1 rows and 1 columns:
\describe{
  \item{mytext}{unsplit list of words containing musical instruments, weapongs, fruits, emotions}
}
}
\usage{
Unordered_List
}
\description{
No talker delineated. List of 17 words spanning 4 semantic categories, Good for examining clustering
}
\keyword{datasets}
