% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_matrix2dendrogram.R
\name{viz_matrix2dendrogram}
\alias{viz_matrix2dendrogram}
\title{viz_matrix2dendrogram}
\usage{
viz_matrix2dendrogram(distmat)
}
\arguments{
\item{distmat}{a square matrix of cosine dist values generated by 'dist_matrix' function}
}
\value{
a plot of a dendrogram
}
\description{
viz_matrix2dendrogram
}
